+++ runtime/calamari/src/migrations/sudo.rs
@@ -38,7 +38,7 @@ impl<T: frame_system::Config> OnRuntimeUpgrade for RemoveSudo<T> {
             log::info!(target: "OnRuntimeUpgrade", "✅ The pallet version has been removed.");
             T::DbWeight::get()
                 .reads(1)
-                .saturating_add(T::DbWeight::get().writes(1 as u64))
+                .saturating_add(T::DbWeight::get().writes(1_u64))
         } else {
             T::DbWeight::get().reads(1)
         }
