+++ runtime/calamari/src/weights/cumulus_pallet_xcmp_queue.rs
@@ -56,14 +56,14 @@ impl<T: frame_system::Config> cumulus_pallet_xcmp_queue::WeightInfo for Substrat
     // Storage: XcmpQueue QueueConfig (r:1 w:1)
     fn set_config_with_u32() -> Weight {
         Weight::from_ref_time(5_779_000)
-            .saturating_add(T::DbWeight::get().reads(1 as u64))
-            .saturating_add(T::DbWeight::get().writes(1 as u64))
+            .saturating_add(T::DbWeight::get().reads(1_u64))
+            .saturating_add(T::DbWeight::get().writes(1_u64))
     }
     // Storage: XcmpQueue QueueConfig (r:1 w:1)
     fn set_config_with_weight() -> Weight {
         Weight::from_ref_time(5_753_000)
-            .saturating_add(T::DbWeight::get().reads(1 as u64))
-            .saturating_add(T::DbWeight::get().writes(1 as u64))
+            .saturating_add(T::DbWeight::get().reads(1_u64))
+            .saturating_add(T::DbWeight::get().writes(1_u64))
     }
 }
 
@@ -72,13 +72,13 @@ impl WeightInfo for () {
     // Storage: XcmpQueue QueueConfig (r:1 w:1)
     fn set_config_with_u32() -> Weight {
         Weight::from_ref_time(5_779_000)
-            .saturating_add(RocksDbWeight::get().reads(1 as u64))
-            .saturating_add(RocksDbWeight::get().writes(1 as u64))
+            .saturating_add(RocksDbWeight::get().reads(1_u64))
+            .saturating_add(RocksDbWeight::get().writes(1_u64))
     }
     // Storage: XcmpQueue QueueConfig (r:1 w:1)
     fn set_config_with_weight() -> Weight {
         Weight::from_ref_time(5_753_000)
-            .saturating_add(RocksDbWeight::get().reads(1 as u64))
-            .saturating_add(RocksDbWeight::get().writes(1 as u64))
+            .saturating_add(RocksDbWeight::get().reads(1_u64))
+            .saturating_add(RocksDbWeight::get().writes(1_u64))
     }
 }
