+++ runtime/calamari/src/weights/pallet_timestamp.rs
@@ -56,8 +56,8 @@ impl<T: frame_system::Config> pallet_timestamp::WeightInfo for SubstrateWeight<T
     // Storage: Timestamp Now (r:1 w:1)
     fn set() -> Weight {
         Weight::from_ref_time(7_464_000)
-            .saturating_add(T::DbWeight::get().reads(1 as u64))
-            .saturating_add(T::DbWeight::get().writes(1 as u64))
+            .saturating_add(T::DbWeight::get().reads(1_u64))
+            .saturating_add(T::DbWeight::get().writes(1_u64))
     }
     fn on_finalize() -> Weight {
         Weight::from_ref_time(2_913_000)
@@ -69,8 +69,8 @@ impl WeightInfo for () {
     // Storage: Timestamp Now (r:1 w:1)
     fn set() -> Weight {
         Weight::from_ref_time(7_464_000)
-            .saturating_add(RocksDbWeight::get().reads(1 as u64))
-            .saturating_add(RocksDbWeight::get().writes(1 as u64))
+            .saturating_add(RocksDbWeight::get().reads(1_u64))
+            .saturating_add(RocksDbWeight::get().writes(1_u64))
     }
     fn on_finalize() -> Weight {
         Weight::from_ref_time(2_913_000)
